add file existence check

+++ install.py
@@ -6,6 +6,7 @@ exclude = [ '*.sw*', '.git', 'install.*' ]
 for f in os.listdir('.'):
     if not any(fnmatch.fnmatch(f, p) for p in exclude):
         path = os.getenv('HOME')+'/'+f
-        os.unlink(path)
+        if os.path.isfile(path) or os.path.isdir(path):
+            os.unlink(path)
         os.symlink(os.path.abspath(f), path)
         print 'create link for %s' % (path)
